+++ src/Aws/CloudFront/CloudFrontClient.php
@@ -64,7 +64,7 @@ use Guzzle\Service\Resource\ResourceIteratorInterface;
  */
 class CloudFrontClient extends AbstractClient
 {
-    const LATEST_API_VERSION = '2013-05-12';
+    const LATEST_API_VERSION = '2013-08-26';
 
     /**
      * Factory method to create a new Amazon CloudFront client using an array of configuration options.