+++ src/Aws/CloudFront/Resources/cloudfront-2013-08-26.php
@@ -15,7 +15,7 @@
  */
 
 return array (
-    'apiVersion' => '2013-05-12',
+    'apiVersion' => '2013-08-26',
     'endpointPrefix' => 'cloudfront',
     'serviceFullName' => 'Amazon CloudFront',
     'serviceAbbreviation' => 'CloudFront',
@@ -68,7 +68,7 @@ return array (
     'operations' => array(
         'CreateCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-05-12/origin-access-identity/cloudfront',
+            'uri' => '/2013-08-26/origin-access-identity/cloudfront',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateCloudFrontOriginAccessIdentityResult',
             'responseType' => 'model',
@@ -76,7 +76,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'CloudFrontOriginAccessIdentityConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-05-12/',
+                        'http://cloudfront.amazonaws.com/doc/2013-08-26/',
                     ),
                 ),
             ),
@@ -121,7 +121,7 @@ return array (
         ),
         'CreateDistribution' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-05-12/distribution',
+            'uri' => '/2013-08-26/distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateDistributionResult',
             'responseType' => 'model',
@@ -129,7 +129,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'DistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-05-12/',
+                        'http://cloudfront.amazonaws.com/doc/2013-08-26/',
                     ),
                 ),
             ),
@@ -412,6 +412,38 @@ return array (
                         ),
                     ),
                 ),
+                'CustomErrorResponses' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'Quantity' => array(
+                            'required' => true,
+                            'type' => 'numeric',
+                        ),
+                        'Items' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'CustomErrorResponse',
+                                'type' => 'object',
+                                'properties' => array(
+                                    'ErrorCode' => array(
+                                        'required' => true,
+                                        'type' => 'numeric',
+                                    ),
+                                    'ResponsePagePath' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'ResponseCode' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'ErrorCachingMinTTL' => array(
+                                        'type' => 'numeric',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
                 'Comment' => array(
                     'required' => true,
                     'type' => 'string',
@@ -524,6 +556,16 @@ return array (
                     'class' => 'InvalidDefaultRootObjectException',
                 ),
                 array(
+                    'reason' => 'The relative path is too big, is not URL-encoded, or does not begin with a slash (/).',
+                    'class' => 'InvalidRelativePathException',
+                ),
+                array(
+                    'class' => 'InvalidErrorCodeException',
+                ),
+                array(
+                    'class' => 'InvalidResponseCodeException',
+                ),
+                array(
                     'reason' => 'The argument is invalid.',
                     'class' => 'InvalidArgumentException',
                 ),
@@ -563,7 +605,7 @@ return array (
         ),
         'CreateInvalidation' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-05-12/distribution/{DistributionId}/invalidation',
+            'uri' => '/2013-08-26/distribution/{DistributionId}/invalidation',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateInvalidationResult',
             'responseType' => 'model',
@@ -571,7 +613,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'InvalidationBatch',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-05-12/',
+                        'http://cloudfront.amazonaws.com/doc/2013-08-26/',
                     ),
                 ),
             ),
@@ -641,7 +683,7 @@ return array (
         ),
         'CreateStreamingDistribution' => array(
             'httpMethod' => 'POST',
-            'uri' => '/2013-05-12/streaming-distribution',
+            'uri' => '/2013-08-26/streaming-distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'CreateStreamingDistributionResult',
             'responseType' => 'model',
@@ -649,7 +691,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'StreamingDistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-05-12/',
+                        'http://cloudfront.amazonaws.com/doc/2013-08-26/',
                     ),
                 ),
             ),
@@ -811,9 +853,9 @@ return array (
         ),
         'DeleteCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'DELETE',
-            'uri' => '/2013-05-12/origin-access-identity/cloudfront/{Id}',
+            'uri' => '/2013-08-26/origin-access-identity/cloudfront/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
-            'responseClass' => 'DeleteCloudFrontOriginAccessIdentity2013_05_12Output',
+            'responseClass' => 'DeleteCloudFrontOriginAccessIdentity2013_08_26Output',
             'responseType' => 'model',
             'parameters' => array(
                 'Id' => array(
@@ -851,9 +893,9 @@ return array (
         ),
         'DeleteDistribution' => array(
             'httpMethod' => 'DELETE',
-            'uri' => '/2013-05-12/distribution/{Id}',
+            'uri' => '/2013-08-26/distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
-            'responseClass' => 'DeleteDistribution2013_05_12Output',
+            'responseClass' => 'DeleteDistribution2013_08_26Output',
             'responseType' => 'model',
             'parameters' => array(
                 'Id' => array(
@@ -891,9 +933,9 @@ return array (
         ),
         'DeleteStreamingDistribution' => array(
             'httpMethod' => 'DELETE',
-            'uri' => '/2013-05-12/streaming-distribution/{Id}',
+            'uri' => '/2013-08-26/streaming-distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
-            'responseClass' => 'DeleteStreamingDistribution2013_05_12Output',
+            'responseClass' => 'DeleteStreamingDistribution2013_08_26Output',
             'responseType' => 'model',
             'parameters' => array(
                 'Id' => array(
@@ -931,7 +973,7 @@ return array (
         ),
         'GetCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/origin-access-identity/cloudfront/{Id}',
+            'uri' => '/2013-08-26/origin-access-identity/cloudfront/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetCloudFrontOriginAccessIdentityResult',
             'responseType' => 'model',
@@ -959,7 +1001,7 @@ return array (
         ),
         'GetCloudFrontOriginAccessIdentityConfig' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/origin-access-identity/cloudfront/{Id}/config',
+            'uri' => '/2013-08-26/origin-access-identity/cloudfront/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetCloudFrontOriginAccessIdentityConfigResult',
             'responseType' => 'model',
@@ -987,7 +1029,7 @@ return array (
         ),
         'GetDistribution' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/distribution/{Id}',
+            'uri' => '/2013-08-26/distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetDistributionResult',
             'responseType' => 'model',
@@ -1015,7 +1057,7 @@ return array (
         ),
         'GetDistributionConfig' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/distribution/{Id}/config',
+            'uri' => '/2013-08-26/distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetDistributionConfigResult',
             'responseType' => 'model',
@@ -1043,7 +1085,7 @@ return array (
         ),
         'GetInvalidation' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/distribution/{DistributionId}/invalidation/{Id}',
+            'uri' => '/2013-08-26/distribution/{DistributionId}/invalidation/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetInvalidationResult',
             'responseType' => 'model',
@@ -1080,7 +1122,7 @@ return array (
         ),
         'GetStreamingDistribution' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/streaming-distribution/{Id}',
+            'uri' => '/2013-08-26/streaming-distribution/{Id}',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetStreamingDistributionResult',
             'responseType' => 'model',
@@ -1108,7 +1150,7 @@ return array (
         ),
         'GetStreamingDistributionConfig' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/streaming-distribution/{Id}/config',
+            'uri' => '/2013-08-26/streaming-distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'GetStreamingDistributionConfigResult',
             'responseType' => 'model',
@@ -1136,7 +1178,7 @@ return array (
         ),
         'ListCloudFrontOriginAccessIdentities' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/origin-access-identity/cloudfront',
+            'uri' => '/2013-08-26/origin-access-identity/cloudfront',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListCloudFrontOriginAccessIdentitiesResult',
             'responseType' => 'model',
@@ -1163,7 +1205,7 @@ return array (
         ),
         'ListDistributions' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/distribution',
+            'uri' => '/2013-08-26/distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListDistributionsResult',
             'responseType' => 'model',
@@ -1190,7 +1232,7 @@ return array (
         ),
         'ListInvalidations' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/distribution/{DistributionId}/invalidation',
+            'uri' => '/2013-08-26/distribution/{DistributionId}/invalidation',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListInvalidationsResult',
             'responseType' => 'model',
@@ -1230,7 +1272,7 @@ return array (
         ),
         'ListStreamingDistributions' => array(
             'httpMethod' => 'GET',
-            'uri' => '/2013-05-12/streaming-distribution',
+            'uri' => '/2013-08-26/streaming-distribution',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'ListStreamingDistributionsResult',
             'responseType' => 'model',
@@ -1257,7 +1299,7 @@ return array (
         ),
         'UpdateCloudFrontOriginAccessIdentity' => array(
             'httpMethod' => 'PUT',
-            'uri' => '/2013-05-12/origin-access-identity/cloudfront/{Id}/config',
+            'uri' => '/2013-08-26/origin-access-identity/cloudfront/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'UpdateCloudFrontOriginAccessIdentityResult',
             'responseType' => 'model',
@@ -1265,7 +1307,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'CloudFrontOriginAccessIdentityConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-05-12/',
+                        'http://cloudfront.amazonaws.com/doc/2013-08-26/',
                     ),
                 ),
             ),
@@ -1332,7 +1374,7 @@ return array (
         ),
         'UpdateDistribution' => array(
             'httpMethod' => 'PUT',
-            'uri' => '/2013-05-12/distribution/{Id}/config',
+            'uri' => '/2013-08-26/distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'UpdateDistributionResult',
             'responseType' => 'model',
@@ -1340,7 +1382,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'DistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-05-12/',
+                        'http://cloudfront.amazonaws.com/doc/2013-08-26/',
                     ),
                 ),
             ),
@@ -1623,6 +1665,38 @@ return array (
                         ),
                     ),
                 ),
+                'CustomErrorResponses' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'Quantity' => array(
+                            'required' => true,
+                            'type' => 'numeric',
+                        ),
+                        'Items' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'CustomErrorResponse',
+                                'type' => 'object',
+                                'properties' => array(
+                                    'ErrorCode' => array(
+                                        'required' => true,
+                                        'type' => 'numeric',
+                                    ),
+                                    'ResponsePagePath' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'ResponseCode' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'ErrorCachingMinTTL' => array(
+                                        'type' => 'numeric',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
                 'Comment' => array(
                     'required' => true,
                     'type' => 'string',
@@ -1734,6 +1808,16 @@ return array (
                     'class' => 'InvalidDefaultRootObjectException',
                 ),
                 array(
+                    'reason' => 'The relative path is too big, is not URL-encoded, or does not begin with a slash (/).',
+                    'class' => 'InvalidRelativePathException',
+                ),
+                array(
+                    'class' => 'InvalidErrorCodeException',
+                ),
+                array(
+                    'class' => 'InvalidResponseCodeException',
+                ),
+                array(
                     'reason' => 'The argument is invalid.',
                     'class' => 'InvalidArgumentException',
                 ),
@@ -1788,7 +1872,7 @@ return array (
         ),
         'UpdateStreamingDistribution' => array(
             'httpMethod' => 'PUT',
-            'uri' => '/2013-05-12/streaming-distribution/{Id}/config',
+            'uri' => '/2013-08-26/streaming-distribution/{Id}/config',
             'class' => 'Guzzle\\Service\\Command\\OperationCommand',
             'responseClass' => 'UpdateStreamingDistributionResult',
             'responseType' => 'model',
@@ -1796,7 +1880,7 @@ return array (
                 'xmlRoot' => array(
                     'name' => 'StreamingDistributionConfig',
                     'namespaces' => array(
-                        'http://cloudfront.amazonaws.com/doc/2013-05-12/',
+                        'http://cloudfront.amazonaws.com/doc/2013-08-26/',
                     ),
                 ),
             ),
@@ -2294,6 +2378,36 @@ return array (
                                 ),
                             ),
                         ),
+                        'CustomErrorResponses' => array(
+                            'type' => 'object',
+                            'properties' => array(
+                                'Quantity' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'Items' => array(
+                                    'type' => 'array',
+                                    'items' => array(
+                                        'name' => 'CustomErrorResponse',
+                                        'type' => 'object',
+                                        'sentAs' => 'CustomErrorResponse',
+                                        'properties' => array(
+                                            'ErrorCode' => array(
+                                                'type' => 'numeric',
+                                            ),
+                                            'ResponsePagePath' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'ResponseCode' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'ErrorCachingMinTTL' => array(
+                                                'type' => 'numeric',
+                                            ),
+                                        ),
+                                    ),
+                                ),
+                            ),
+                        ),
                         'Comment' => array(
                             'type' => 'string',
                         ),
@@ -2551,7 +2665,7 @@ return array (
                 ),
             ),
         ),
-        'DeleteCloudFrontOriginAccessIdentity2013_05_12Output' => array(
+        'DeleteCloudFrontOriginAccessIdentity2013_08_26Output' => array(
             'type' => 'object',
             'additionalProperties' => true,
             'properties' => array(
@@ -2561,7 +2675,7 @@ return array (
                 ),
             ),
         ),
-        'DeleteDistribution2013_05_12Output' => array(
+        'DeleteDistribution2013_08_26Output' => array(
             'type' => 'object',
             'additionalProperties' => true,
             'properties' => array(
@@ -2571,7 +2685,7 @@ return array (
                 ),
             ),
         ),
-        'DeleteStreamingDistribution2013_05_12Output' => array(
+        'DeleteStreamingDistribution2013_08_26Output' => array(
             'type' => 'object',
             'additionalProperties' => true,
             'properties' => array(
@@ -2920,6 +3034,36 @@ return array (
                                 ),
                             ),
                         ),
+                        'CustomErrorResponses' => array(
+                            'type' => 'object',
+                            'properties' => array(
+                                'Quantity' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'Items' => array(
+                                    'type' => 'array',
+                                    'items' => array(
+                                        'name' => 'CustomErrorResponse',
+                                        'type' => 'object',
+                                        'sentAs' => 'CustomErrorResponse',
+                                        'properties' => array(
+                                            'ErrorCode' => array(
+                                                'type' => 'numeric',
+                                            ),
+                                            'ResponsePagePath' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'ResponseCode' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'ErrorCachingMinTTL' => array(
+                                                'type' => 'numeric',
+                                            ),
+                                        ),
+                                    ),
+                                ),
+                            ),
+                        ),
                         'Comment' => array(
                             'type' => 'string',
                         ),
@@ -3186,6 +3330,37 @@ return array (
                                         'type' => 'string',
                                     ),
                                     'MinTTL' => array(
+                                        'type' => 'numeric',
+                                    ),
+                                ),
+                            ),
+                        ),
+                    ),
+                ),
+                'CustomErrorResponses' => array(
+                    'type' => 'object',
+                    'location' => 'xml',
+                    'properties' => array(
+                        'Quantity' => array(
+                            'type' => 'numeric',
+                        ),
+                        'Items' => array(
+                            'type' => 'array',
+                            'items' => array(
+                                'name' => 'CustomErrorResponse',
+                                'type' => 'object',
+                                'sentAs' => 'CustomErrorResponse',
+                                'properties' => array(
+                                    'ErrorCode' => array(
+                                        'type' => 'numeric',
+                                    ),
+                                    'ResponsePagePath' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'ResponseCode' => array(
+                                        'type' => 'string',
+                                    ),
+                                    'ErrorCachingMinTTL' => array(
                                         'type' => 'numeric',
                                     ),
                                 ),
@@ -3837,6 +4012,36 @@ return array (
                                     ),
                                 ),
                             ),
+                            'CustomErrorResponses' => array(
+                                'type' => 'object',
+                                'properties' => array(
+                                    'Quantity' => array(
+                                        'type' => 'numeric',
+                                    ),
+                                    'Items' => array(
+                                        'type' => 'array',
+                                        'items' => array(
+                                            'name' => 'CustomErrorResponse',
+                                            'type' => 'object',
+                                            'sentAs' => 'CustomErrorResponse',
+                                            'properties' => array(
+                                                'ErrorCode' => array(
+                                                    'type' => 'numeric',
+                                                ),
+                                                'ResponsePagePath' => array(
+                                                    'type' => 'string',
+                                                ),
+                                                'ResponseCode' => array(
+                                                    'type' => 'string',
+                                                ),
+                                                'ErrorCachingMinTTL' => array(
+                                                    'type' => 'numeric',
+                                                ),
+                                            ),
+                                        ),
+                                    ),
+                                ),
+                            ),
                             'Comment' => array(
                                 'type' => 'string',
                             ),
@@ -4334,6 +4539,36 @@ return array (
                                                 'type' => 'string',
                                             ),
                                             'MinTTL' => array(
+                                                'type' => 'numeric',
+                                            ),
+                                        ),
+                                    ),
+                                ),
+                            ),
+                        ),
+                        'CustomErrorResponses' => array(
+                            'type' => 'object',
+                            'properties' => array(
+                                'Quantity' => array(
+                                    'type' => 'numeric',
+                                ),
+                                'Items' => array(
+                                    'type' => 'array',
+                                    'items' => array(
+                                        'name' => 'CustomErrorResponse',
+                                        'type' => 'object',
+                                        'sentAs' => 'CustomErrorResponse',
+                                        'properties' => array(
+                                            'ErrorCode' => array(
+                                                'type' => 'numeric',
+                                            ),
+                                            'ResponsePagePath' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'ResponseCode' => array(
+                                                'type' => 'string',
+                                            ),
+                                            'ErrorCachingMinTTL' => array(
                                                 'type' => 'numeric',
                                             ),
                                         ),